+++ tests/testthat/test-examples.R
@@ -11,3 +11,12 @@ test_with_dir("examples are listed and written", {
     unlink(i, recursive = TRUE, force = TRUE)
   }
 })
+
+test_with_dir("overwrites of report.Rmd handled correctly", {
+  expect_silent(load_basic_example(overwrite = TRUE))
+  expect_silent(load_basic_example(overwrite = FALSE))
+  expect_warning(load_basic_example(overwrite = TRUE))
+  expect_warning(load_basic_example(to = "a", report_file = "b"))
+  expect_true(file.exists("b"))
+  expect_false(file.exists("a"))
+})