+++ tests/testthat/test-parallel-ui.R
@@ -12,6 +12,7 @@ test_with_dir("shell_file() writes correctly", {
   shell_file()
   expect_true(file.exists("shell.sh"))
   unlink("shell.sh", force = TRUE)
+
   d <- "exdir"
   dir.create(d)
   p <- file.path(d, "script.txt")
@@ -19,6 +20,10 @@ test_with_dir("shell_file() writes correctly", {
   shell_file(p)
   expect_true(file.exists(p))
   unlink(d, recursive = TRUE, force = TRUE)
+
+  expect_silent(shell_file(overwrite = TRUE))
+  expect_silent(shell_file(overwrite = FALSE))
+  expect_warning(shell_file(overwrite = TRUE))
 })
 
 test_with_dir("mclapply and lapply", {